Draw disabled X-Ray button when necessary
I just remembered that you can make a row look like it belongs in a pie menu by setting emboss and scale, so instead of having a blank in the pie when X-Ray is not possible, it now draws it as disabled.
There does seem to be a weird bug where you can still enable it if you press and hold Z, hover your mouse exactly over the button, then release Z, but whatever.