Add view utils
Just one function at the moment, closest_axis, which finds the closest axis to the current view, or the opposite of that if flip=True.
Notice how there's spaces around the = sign? Apparently that's how you do it for function definitions, but only if the parameter in question is type hinted. So if the default value is None, you would not type hint it, and not add spaces around the = sign.
Odd but whatever.