Make keymaps not error when Blender is launched in the background
When Blender is launched as a background process, register fails because the keyconfig isn't ready at the time of register.
Because register fails, unregister also fails, since there is no keymap to remove.
I now only do anything with the keyconfig if it exists, and only do register the keymap if it doesn't exist, and only unregister the keymap if it does exist.
I figure this means you won't be able to use the shortcuts while Blender is running in the background, but it doesn't matter because... it's running in the background.