Prevent other addon keymap items from showing up
BoxCutter for example has a keymap item in the same keymap this addon uses, which then shows up in the Alternative Pies preferences.
This is because keymaps.new doesn't necessarily create a unique keymap, but can instead return an existing one if the name and space type match.
The solution is to not rely on keymap.keymap_items for getting our list of keymap items, but instead keep a separate list and add the items to it manually.